made UFO a random number of points

+++ SpaceInvaders/Gift.pde
@@ -1,6 +1,6 @@
 public class Gift extends Alien {  
   
-  Gift(int x, int y, int s, PImage i) { // s parameter is speed
+  Gift(int x, int y, float s, PImage i) { // s parameter is speed
     super(x, y, 0, i, i); // last parameter is score, it = 0 because gifts don't add score
     this.xDiameter = 25;
     this.yDiameter = 25;
